Split the tests apart into multiple files
